utils_3.hに関数プロトタイプを分割
