+++ Makefile
@@ -6,7 +6,7 @@
 #    By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/04/03 12:55:20 by ttsubo            #+#    #+#              #
-#    Updated: 2025/04/28 14:52:35 by ttsubo           ###   ########.fr        #
+#    Updated: 2025/04/28 15:48:10 by ttsubo           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -53,7 +53,7 @@ TOKENIZER_SRC	= 	tokenizer.c tokenizer_error.c read_token.c \
 PARSER_SRC		=	allocate_cmds.c  parser.c  parser_utils.c  setup_cmds.c \
 					expand_env.c	expand_env_utils.c
 BUILTIN_SRC		=	cd.c exit.c pwd.c echo.c env.c unset.c export.c \
-					env_utils.c env_utils_2.c builtin_utils.c export_print_sorted_env.c
+					env_utils.c env_utils_2.c env_utils_3.c builtin_utils.c export_print_sorted_env.c
 
 # 3.Add more directories as they are added.
 $(eval $(call add_module,root))
